Add simple inventory dumper to executables. Move filereader to independent module so as to not rely on XRootD presence.
