+++ bin/consistency-dump-tree
@@ -1,0 +1,16 @@
+#! /usr/bin/env python
+
+"""
+Dumps the inventory tree into ``<VarDirectory>/cache``
+"""
+
+__usage__ = '%prog [options]'
+
+from dynamo_consistency import inventorylister
+from dynamo_consistency import opts
+from dynamo_consistency.parser import pretty_exe
+
+pretty_exe('consistency-dump-tree')
+
+if __name__ == '__main__':
+    inventorylister.listing(opts.SITE_PATTERN)
