+++ docs/consistency_dump_tree_exe.py
@@ -1,0 +1,1 @@
+../bin/consistency-dump-tree
