+++ dynamo_consistency/backend/listers.py
@@ -6,7 +6,6 @@ Module containing the classes that are used for remote listing
 """
 
 import logging
-import os
 import re
 import time
 import subprocess
@@ -16,7 +15,6 @@ import XRootD.client    # pylint: disable=import-error
 
 from . import redirectors
 from .. import config
-from .. import datatypes
 
 LOG = logging.getLogger(__name__)
 
@@ -316,41 +314,6 @@ class NoPath(Exception):
 GFAL_LOCATION = lambda _: ''
 
 
-def file_reader(filename, translator):
-    """
-    :param str filename: Name of a file to read file information from
-    :param func translator: A function that takes a line of a file as an arument
-                            and transforms the line into input for
-                            :py:func:`dynamo_consistencydatatypes.DirectoryInfo.add_file_list`.
-    :returns: A directory tree created from reading the input file
-    :rtype: dynamo_consistency.datatypes.DirectoryInfo
-    """
-
-    tree = datatypes.DirectoryInfo(
-        name=config.config_dict()['RootPath']
-        )
-
-    config_dict = config.config_dict()
-    dirlist = ['%s' % os.path.join(config_dict['RootPath'], directory)
-               for directory in config_dict['DirectoryList']]
-
-    def line_yielder():
-        """
-        Yields the translated lines of the input file
-        for :py:func:`dynamo_consistency.datatypes.DirectoryInfo.add_file_list`
-        """
-        with open(filename, 'r') as inputfile:
-            for line in inputfile:
-                for directory in dirlist:
-                    if line.startswith(directory):
-                        yield translator(line)
-                        break
-
-    tree.add_file_list(line_yielder())
-
-    return tree
-
-
 def get_listers(site):
     """
     Picks out a suitable lister for a site based on the configuration file.
@@ -368,6 +331,7 @@ def get_listers(site):
     access = config_dict.get('AccessMethod', {}).get(site)
 
     if access == 'RAL-Reader':
+        from .filereader import file_reader
         from ..cms.filedumps import read_ral_dump
         return file_reader(*read_ral_dump(GFAL_LOCATION(site)))
